+++ template/jsx/src/pages/index.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import { Index } from "..";
+
+const IndexComp = () => (
+    <Index />
+);
+
+export default IndexComp;